refactor: 제스처 로직 분리 및 zIndex 상수 관리
