+++ src/pages/lanternDetail/constants/zIndex.ts
@@ -1,0 +1,7 @@
+export const Z_INDEX = {
+  OVERLAY: 0,
+  SCROLL_CONTAINER: 1,
+  CLOSE_BUTTON: 1000,
+  HAND_POINTER: 2000,
+  INTERACTION_MESSAGE: 3000,
+} as const;
